pybind conversion for IntList
also fix #11204 to use the utility function
