+++ torch/csrc/utils/pybind.h
@@ -8,6 +8,8 @@
 
 #include "torch/csrc/DynamicTypes.h"
 #include "torch/csrc/autograd/python_variable.h"
+#include "torch/csrc/utils/python_tuples.h"
+#include "torch/csrc/utils/python_numbers.h"
 
 #include <stdexcept>
 
@@ -55,6 +57,39 @@ public:
   static handle cast(torch::autograd::Variable src, return_value_policy /* policy */, handle /* parent */) {
     return handle(THPVariable_Wrap(src));
   }
+};
+
+template<> struct type_caster<at::IntList> {
+public:
+  PYBIND11_TYPE_CASTER(at::IntList, _("at::IntList"));
+
+  bool load(handle src, bool) {
+    PyObject *source = src.ptr();
+    auto tuple = PyTuple_Check(source);
+    if (tuple || PyList_Check(source)) {
+      auto size = tuple ? PyTuple_GET_SIZE(source) : PyList_GET_SIZE(source);
+      v_value.resize(size);
+      for (int idx = 0; idx < size; idx++) {
+	PyObject* obj = tuple ? PyTuple_GET_ITEM(source, idx) : PyList_GET_ITEM(source, idx);
+	if (THPVariable_Check(obj)) {
+	  v_value[idx] = THPVariable_Unpack(obj).toCLong();
+	} else if (PyLong_Check(obj)) {
+	  // use THPUtils_unpackLong after it is safe to include python_numbers.h
+	  v_value[idx] = THPUtils_unpackLong(obj);
+	} else {
+	  return false;
+	}
+      }
+      value = v_value;
+      return true;
+    }
+    return false;
+  }
+  static handle cast(at::IntList src, return_value_policy /* policy */, handle /* parent */) {
+    return handle(THPUtils_packInt64Array(src.size(), src.data()));
+  }
+private:
+  std::vector<int64_t> v_value;
 };
 
 // http://pybind11.readthedocs.io/en/stable/advanced/cast/stl.html#c-17-library-containers
